// src/types/woocommerce.d.ts
interface LineItem {
	id?: number;
	name?: string;
	product_id: number;
	variation_id?: number;
	quantity: number;
	tax_class?: string;
	subtotal?: string;
	subtotal_tax?: string;
	total?: string;
	total_tax?: string;
	taxes?: any[];
	meta_data?: MetaData[];
	sku?: string;
	price?: number;
}
// ...
export interface MetaData {
	id: number;
	key: string;
	value: string;
}